# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API REST Panel de Control de Club Deportivo",
        "description": "API para autenticación y gestión de un club deportivo",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/v1/sports": {
            "get": {
                "tags": [
                    "Deportes"
                ],
                "summary": "Obtener lista de deportes",
                "description": "Devuelve un listado paginado de deportes junto con sus canchas asociadas",
                "operationId": "59bc921dc5d60ecd9fbdc9995d7ca6b9",
                "responses": {
                    "200": {
                        "description": "Lista de deportes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SportResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Deportes"
                ],
                "summary": "Crear un nuevo deporte",
                "description": "Registra un nuevo deporte en la base de datos",
                "operationId": "e82e4b406d69af25a836690663a9e946",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreSportRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Deporte creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SportResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/sports/{id}": {
            "get": {
                "tags": [
                    "Deportes"
                ],
                "summary": "Obtener un deporte por ID",
                "description": "Devuelve los datos de un deporte específico junto con sus canchas asociadas",
                "operationId": "46bf7bdbd39dfc8829aeaf06d47c3fda",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del deporte",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deporte encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SportResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Deporte no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Deportes"
                ],
                "summary": "Actualizar un deporte",
                "description": "Modifica un deporte existente en la base de datos",
                "operationId": "13826660186b7738eb292b4978456ec0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del deporte",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSportRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deporte actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SportResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Deporte no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Deportes"
                ],
                "summary": "Eliminar un deporte",
                "description": "Borra un deporte usando su ID",
                "operationId": "d4ea7921ed67b57885f619b27524dde6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del deporte",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deporte eliminado correctamente"
                    },
                    "404": {
                        "description": "Deporte no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Deportes"
                ],
                "summary": "Actualizar parcialmente un deporte",
                "description": "Modifica solo los campos enviados en la solicitud",
                "operationId": "4147df603d73e4223b2544862fbb3256",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del deporte",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSportRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deporte actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SportResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Deporte no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreSportRequest": {
                "title": "Store Sport Request",
                "description": "Datos necesarios para registrar un nuevo deporte",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Fútbol"
                    },
                    "description": {
                        "type": "string",
                        "example": "Deporte en el que dos equipos compiten para anotar goles"
                    }
                },
                "type": "object"
            },
...
            "UpdateSportRequest": {
                "title": "Update Sport Request",
                "description": "Datos necesarios para actualizar un deporte",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Fútbol"
                    },
                    "description": {
                        "type": "string",
                        "example": "Deporte en el que dos equipos compiten para anotar goles"
                    }
                },
                "type": "object"
            },
...
            "SportResource": {
                "title": "Sport Resource",
                "description": "Representación de un deporte en la API",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Tenis"
                    },
                    "description": {
                        "type": "string",
                        "example": "Deporte de raqueta jugado en una cancha rectangular."
                    },
                    "courts": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 3
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Court A"
                                },
                                "sport_id": {
                                    "type": "integer",
                                    "example": 1
                                }
                            },
                            "type": "object"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-14T12:05:09Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-14T12:10:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}